+++ cleo_plugins/Audio/Audio.cpp
@@ -94,7 +94,7 @@ public:
     //0AAC=2,  %2d% = load_audiostream %1d%  // IF and SET
     static OpcodeResult __stdcall opcode_0AAC(CScriptThread* thread)
     {
-        auto path = OPCODE_READ_PARAM_FILEPATH();
+        OPCODE_READ_PARAM_FILEPATH(path);
 
         auto ptr = soundSystem.CreateStream(path);
 
@@ -196,7 +196,7 @@ public:
     //0AC1=2,%2d% = load_audiostream_with_3d_support %1d% //IF and SET
     static OpcodeResult __stdcall opcode_0AC1(CScriptThread* thread)
     {
-        auto path = OPCODE_READ_PARAM_FILEPATH();
+        OPCODE_READ_PARAM_FILEPATH(path);
 
         auto ptr = soundSystem.CreateStream(path, true);